requirements.txt
fastapi~=0.115.5
uvicorn
sqlalchemy~=2.0.36
psycopg2-binary
pydantic~=2.10.2
passlib~=1.7.4
jwt~=1.3.1
python-dotenv~=1.0.1